+++ __init__.py
@@ -3,7 +3,7 @@
 bl_info = {
     "name": "Path Actions",
     "author": "Samuel Bernou",
-    "version": (2, 4, 0),
+    "version": (2, 4, 1),
     "blender": (4, 2, 0),
     "location": "Window top right corner, browser up-left corner, addon prefs",
     "description": "Open blend folder in OS explorer",
@@ -15,6 +15,7 @@ bl_info = {
 from . import (
             operators,
             history,
+            links_checker,
             blend_open_ops,
             extend_history,
             open_addons_path,
@@ -26,6 +27,7 @@ mods = (
     operators,
     blend_open_ops,
     history,
+    links_checker,
     extend_history,
     open_addons_path,
     panels,